Update Sources' README
Improve the "Output Docs" section by adding descriptions of the markdown
include-files, and document the difference between the HTML docs in the
`docs-src/` folder and those in the root (i.e. the devs' previews and
the current releases).

Also add some ref-links labels for project files, to reduce verbosity
in the source document.